Move pipeline mocks out of jest.setup.ts
jest.setup.ts was globally mocking both generate_csv/generate and
generate_json/generate, which meant no test could ever exercise the
real pipelines unless it explicitly un-mocked them. The only test
that actually needs those stubs is translate.spec.ts, which exercises
the orchestration around the pipelines rather than the pipelines
themselves.

Hoist the mocks into that file, clean up the matching jest.unmock
calls in concurrency.spec.ts, and leave jest.setup.ts with just the
ChatFactory and utils mocks that multiple suites share.

No behaviour change; 82 tests still pass.
